+++ NEWS.md
@@ -1,8 +1,9 @@
-# aqp 1.43 (2022-04-11)
+# aqp 1.43 (2022-07-14)
  * `plotSPC()` gains argument `maxLabelAdjustmentIndex` for controlling horizon depth label collisions
  * bug fix in `plotSPC()` when `fixLabelCollisions = TRUE`, adjustments suggested to `fixOverlap()` are now scaled correctly
  * `explainPlotSPC()` reports label adjustment index when label collision repair is enabled
  * aesthetic cleanup in `explainPlotSPC()`
+ * `soilColorSignature()` gains arguments and perceptual color distances (dE00) via farver package
 
 # aqp 1.42 (2022-04-08)
  * CRAN release